Handle reMarkable 3.14 not including size/page count in directory
Some time between 3.10 and 3.14, reMarkable changed the JSON structure
returned by the `/documents/...` endpoint so it doesn't include each
document's size or page count. This updates the program to not try and
parse this information from the JSON if it isn't there, and makes the
`rmweb list` command not show anything in those columns in its output.